feat: inline js_util
